requirements.txt
absl-py==1.4.0
anyio==3.6.2
argcomplete==1.10.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.8.2
bleach==6.0.0
blis==0.7.9
bs4==0.0.1
cachetools==5.3.0
Camelot==12.6.29
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
compressed-rtf==1.0.6
confection==0.0.4
contourpy==1.0.7
Counter==1.0.0
cycler==0.11.0
cymem==2.0.7
Cython==0.29.32
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.8.0
docx==0.2.4
docx2pdf==0.1.8
docx2txt==0.8
ebcdic==1.1.1
Elixir==0.7.1
et-xmlfile==1.1.0
executing==1.2.0
extract-msg==0.28.7
fastapi==0.92.0
fastjsonschema==2.16.2
filelock==3.10.2
fire==0.5.0
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
flatbuffers==23.1.21
fonttools==4.38.0
fqdn==1.5.1
fst-pso==1.8.1
FuzzyTM==2.0.5
gast==0.4.0
gensim==4.3.0
google-auth==2.16.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.3
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
h5py==3.8.0
hpack==4.0.0
html5lib==1.1
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
imageio==2.27.0
IMAPClient==2.1.0
imbalanced-learn==0.10.1
imblearn==0.0
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.6
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.5.1
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keras==2.12.0
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.2
libclang==15.0.6.1
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
miniful==0.0.6
mistune==2.0.5
mpmath==1.2.1
murmurhash==1.0.9
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
oauthlib==3.2.2
olefile==0.46
opencv-python==4.7.0.72
openpyxl==3.1.1
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pathy==0.10.1
pbr==5.11.1
pdf2docx==0.5.6
pdfminer==20191125
pdfminer.six==20191110
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
polling2==0.5.0
portalocker==2.7.0
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==4.22.0
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.5
pyFUME==0.2.25
Pygments==2.14.0
PyMuPDF==1.21.1
pyparsing==3.0.9
pypdf==3.4.1
PyPDF2==2.12.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
python-json-logger==2.0.6
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.1
scipy==1.10.1
Send2Trash==1.8.0
simpful==2.9.0
six==1.12.0
sklearn==0.0.post1
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
speechmatics-python==1.6.4
SpeechRecognition==3.8.1
SQLAlchemy==2.0.7
sqlalchemy-migrate==0.11.0
sqlparse==0.4.3
srsly==2.4.6
stack-data==0.6.2
starlette==0.25.0
sympy==1.11.1
tabula==1.0.5
tabula-py==2.6.0
Tempita==0.5.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.17.1
textract==1.6.5
thinc==8.1.9
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.2.1
toml==0.10.2
torch==2.0.0+cu117
torchaudio==2.0.1+cu117
torchdata==0.6.0
torchtext==0.15.0
torchvision==0.15.1+cu117
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uri-template==1.2.0
urllib3==1.26.14
utils==1.0.1
uvicorn==0.20.0
virtualenv==20.21.0
wasabi==1.1.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wrapt==1.14.1
xgboost==1.7.4
xlrd==1.2.0
XlsxWriter==3.0.8
xlwt==0.7.2
